pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "creatory"
version = "0.1.0"
description = "Creator-first multi-agent orchestration backend"
readme = "README.md"
requires-python = ">=3.12"
authors = [
  { name = "Creatory Contributors" }
]
dependencies = [
  "fastapi>=0.116.1,<1.0.0",
  "uvicorn[standard]>=0.35.0,<1.0.0",
  "pydantic>=2.11.7,<3.0.0",
  "pydantic-settings>=2.10.1,<3.0.0",
  "sqlalchemy>=2.0.41,<3.0.0",
  "alembic>=1.16.4,<2.0.0",
  "asyncpg>=0.30.0,<1.0.0",
  "gunicorn>=23.0.0,<24.0.0",
  "python-jose[cryptography]>=3.5.0,<4.0.0",
  "passlib[bcrypt]>=1.7.4,<2.0.0",
  "bcrypt==4.0.1",
  "PyYAML>=6.0.2,<7.0.0",
  "python-multipart>=0.0.20,<1.0.0",
  "email-validator>=2.2.0,<3.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.1,<9.0.0",
  "pytest-asyncio>=1.1.0,<2.0.0",
  "httpx>=0.28.1,<1.0.0",
  "ruff>=0.12.5,<1.0.0",
  "black>=25.1.0,<26.0.0",
  "pre-commit>=4.2.0,<5.0.0"
]

[tool.setuptools.packages.find]
include = ["creatory_core*"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]
